+++ src/main/java/me/piggypiglet/randomspawn/data/spawn/types/RadiusSpawn.java
@@ -9,20 +9,20 @@ import org.bukkit.World;
 // https://www.piggypiglet.me
 // ------------------------------
 public final class RadiusSpawn extends Spawn {
-    private double[] center;
+    private int[] center;
     private int radius;
 
-    public RadiusSpawn(String name, Spawns type, String permission, boolean enabled, World world, Options options, double[] center, int radius) {
+    public RadiusSpawn(String name, Spawns type, String permission, boolean enabled, World world, Options options, int[] center, int radius) {
         super(name, type, permission, enabled, world, options);
         this.center = center;
         this.radius = radius;
     }
 
-    public double[] getCenter() {
+    public int[] getCenter() {
         return center;
     }
 
-    public void setCenter(double[] center) {
+    public void setCenter(int[] center) {
         this.center = center;
     }
